requirements/common.txt
dj-database-url==0.4.1
Django==1.9.6
django-bootstrap3==7.0.1
django-compressor==2.0
django-compressor-autoprefixer==0.1.0
django-configurations==1.0
django-debug-toolbar==1.4
django-dotenv==1.4.1
django-extensions==1.6.7
django-hstore==1.4.2
django-libsass==0.7
django-widget-tweaks==1.4.1
et-xmlfile==1.0.1
jdcal==1.2
numpy==1.11.1
openpyxl==2.3.5
pandas==0.18.1
Pillow==3.3.0
psycopg2==2.6.1
python-dateutil==2.5.3
pytz==2016.4
PyYAML==3.11
scipy==0.18.0
six==1.10.0
sqlparse==0.1.19
xlrd==1.0.0
xlwt==1.1.2